Migrate gameLayer code in to Level code
